requirements.txt
streamlit
pandas
matplotlib
seaborn
plotly
statsmodels
openpyxl
Pillow
